requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
wikipedia==1.4.0
pyttsx3==2.90
reportlab==4.0.4
